Rename `update_processing_context.py` to `user_context.py`
